@@ -43,85 +38,68 @@
   },
   "dependencies": {
     "abab": "^2.0.0",
-    "file-type": "^8.1.0",
+    "axios": "^0.18.0",
+    "debug": "^4.1.1",
+    "eventemitter3": "^3.1.2",
+    "file-type": "^10.0.0",
     "filestack-loader": "^3.0.4",
-    "is-svg": "^3.0.0",
-    "isutf8": "^2.0.2",
+    "isutf8": "^2.0.3",
     "jsonschema": "^1.2.4",
+    "p-queue": "^4.0.0",
+    "prettier-tslint": "^0.4.2",
     "spark-md5": "^3.0.0",
-    "superagent": "^3.8.3",
-    "tcomb-validation": "^3.4.1",
+    "ts-node": "^8.1.0",
     "tslib": "^1.9.3"
   },
   "devDependencies": {
-    "@babel/core": "^7.3.4",
-    "@babel/preset-env": "^7.3.4",
-    "@types/mime": "^2.0.0",
-    "@types/mocha": "^2.2.48",
-    "@types/node": "^10.5.2",
-    "@types/sinon": "^5.0.1",
-    "@types/spark-md5": "^3.0.0",
-    "@types/superagent": "^3.8.3",
-    "aws-sdk": "^2.291.0",
-    "browserify": "^14.5.0",
-    "connect": "^3.6.6",
-    "connect-prism": "^1.0.20",
-    "cors": "^2.8.4",
-    "del": "^3.0.0",
-    "dotenv": "^4.0.0",
-    "electron": "^3.0.6",
+    "@babel/core": "^7.4.4",
+    "@babel/preset-env": "^7.4.4",
+    "@purtuga/esm-webpack-plugin": "^1.0.1",
+    "@types/jest": "^24.0.12",
+    "@types/mime": "^2.0.1",
+    "@types/node": "^12.0.0",
+    "@types/spark-md5": "^3.0.1",
+    "@types/uuid": "^3.4.4",
+    "aws-sdk": "^2.451.0",
+    "babel-loader": "^8.0.5",
+    "clean-webpack-plugin": "^2.0.2",
+    "codecov": "^3.3.0",
+    "compression-webpack-plugin": "^2.0.0",
+    "del": "^4.1.1",
+    "dotenv": "^8.0.0",
     "envify": "^4.1.0",
-    "express": "^4.16.4",
-    "fs-jetpack": "^2.2.0",
-    "gh-pages": "^1.2.0",
-    "gulp": "^4.0.0",
-    "gulp-better-rollup": "3.1.0",
+    "fs-jetpack": "^2.2.2",
+    "gh-pages": "^2.0.1",
+    "gulp": "^4.0.2",
     "gulp-gzip": "^1.4.2",
     "gulp-rename": "^1.4.0",
     "gulp-replace": "^1.0.0",
-    "gulp-rollup": "^2.16.2",
-    "gulp-sourcemaps": "^2.6.4",
-    "gulp-typescript": "^4.0.2",
-    "gulp-uglify": "^3.0.1",
-    "har-validator": "^5.1.3",
-    "js-yaml": "^3.12.0",
-    "jsdom": "13.0.0",
-    "jsdom-global": "3.0.2",
-    "karma": "^3.1.1",
-    "karma-browserify": "^5.3.0",
-    "karma-browserstack-launcher": "^1.4.0",
-    "karma-chrome-launcher": "^2.2.0",
-    "karma-electron": "^6.0.0",
-    "karma-firefox-launcher": "^1.1.0",
-    "karma-mocha": "^1.3.0",
-    "karma-rollup-preprocessor": "^6.1.0",
-    "mime-types": "^2.1.19",
-    "mocha": "^5.2.0",
-    "npm-run-all": "^4.1.5",
-    "nyc": "^13.3.0",
-    "opn-cli": "^3.1.0",
-    "rollup": "^0.57.0",
-    "rollup-plugin-alias": "^1.4.0",
-    "rollup-plugin-babel": "^4.3.2",
-    "rollup-plugin-commonjs": "^9.2.0",
-    "rollup-plugin-json": "^3.1.0",
-    "rollup-plugin-node-builtins": "^2.1.2",
-    "rollup-plugin-node-globals": "^1.4.0",
-    "rollup-plugin-node-resolve": "^3.4.0",
-    "rollup-plugin-replace": "^2.1.0",
-    "rollup-watch": "^4.3.1",
-    "sinon": "^7.1.1",
-    "standard-version": "^4.4.0",
-    "toxy": "^0.3.16",
-    "trash-cli": "^1.4.0",
-    "tslint": "^5.11.0",
-    "tslint-config-semistandard": "^7.0.0",
+    "gulp-typescript": "^5.0.1",
+    "jest": "^24.8.0",
+    "jest-localstorage-mock": "^2.4.0",
+    "jest-mock": "^24.8.0",
+    "js-yaml": "^3.13.1",
+    "jsdom": "^15.0.0",
+    "lodash.clonedeep": "^4.5.0",
+    "lodash.merge": "^4.6.1",
+    "nock": "^10.0.6",
+    "opn-cli": "^4.1.0",
+    "prettier": "^1.17.0",
+    "prettier-tslint": "^0.4.2",
+    "standard-version": "^6.0.1",
+    "trash-cli": "^2.0.0",
+    "ts-jest": "^24.0.2",
+    "tslint": "^5.16.0",
+    "tslint-config-semistandard": "^8.0.0",
     "typedoc": "^0.14.2",
-    "typescript": "^2.9.2",
-    "uglify-es": "^3.3.10",
-    "url": "^0.11.0",
+    "typescript": "^3.4.5",
     "validate-commit-msg": "^2.14.0",
-    "watchify": "^3.11.1"
+    "webpack": "^4.30.0",
+    "webpack-assets-manifest": "^3.1.1",
+    "webpack-bundle-analyzer": "^3.3.2",
+    "webpack-cli": "^3.3.1",
+    "webpack-node-externals": "^1.7.2",
+    "webpack-subresource-integrity": "^1.3.2"
   },
   "keywords": [
     "filestack",
