@@ -15,9 +15,10 @@
  * limitations under the License.
  */
 
-import * as t from 'tcomb-validation';
 import { Security, Session } from '../client';
-import { checkOptions, removeEmpty } from '../utils';
+import { removeEmpty } from '../utils';
+import { getValidator, PreviewParamsSchema } from './../../schema';
+import { FilestackError, FilestackErrorType } from './../../filestack_error';
 
 export interface PreviewOptions {
   /**
@@ -63,6 +64,7 @@ export const getUrl = (
   if (hasSecurity) {
     baseUrl.push(`security=policy:${policy},signature:${signature}`);
   }
+
   baseUrl.push(handle);
   return baseUrl.join('/');
 };
@@ -79,17 +81,17 @@ export const preview = (session: Session, handle?: string, opts?: PreviewOptions
   if (!handle || typeof handle !== 'string') {
     throw new Error('A valid Filestack handle or storage alias is required for preview');
   }
-  const allowed = [
-    { name: 'id', type: t.String },
-    { name: 'css', type: t.String },
-  ];
 
-  checkOptions('preview', allowed, opts);
+  const validateRes = getValidator(PreviewParamsSchema)(opts);
 
-  const options = removeEmpty(opts);
+  if (validateRes.errors.length) {
+    throw new FilestackError(`Invalid preview params`, validateRes.errors, FilestackErrorType.VALIDATION);
+  }
+
+  const options = removeEmpty(opts || {});
   const url = getUrl(session, handle, options);
 
-  if (options.id) {
+  if (options && options.id) {
     const id = options.id;
     const iframe = document.createElement('iframe');
     const domElement = document.getElementById(id);
