@@ -15,42 +15,10 @@
  * limitations under the License.
  */
 
-/**
- * @private
- */
-export interface FileObj extends File {
-  buffer: Buffer;
-  name: string;
-  size: number;
-  type: string;
-}
-
-/**
- * @private
- */
-export interface PartObj {
-  buffer: any;
-  chunks: any[];
-  chunkSize: number;
-  intelligentOverride: boolean;
-  loaded: number;
-  number: number;
-  request: any;
-  size: number;
-  md5?: string;
-  offset?: number;
-}
+import { StoreBaseParams } from 'src/lib/filelink';
+import { SanitizeOptions } from './../../utils/index';
 
 export interface UploadOptions {
-  host?: string;
-  /**
-   * Node only. Treat the file argument as a path string.
-   */
-  path?: boolean;
-  /**
-   * Set the MIME type of the uploaded file.
-   */
-  mimetype?: string;
   /**
    * Maximum size for file slices. Is overridden when intelligent=true. Default is `6 * 1024 * 1024` (6MB).
    */
@@ -70,7 +38,7 @@ export interface UploadOptions {
   /**
    * Callback for retry events.
    */
-  onRetry?: (evt: FSRetryEvent) => void;
+  onRetry?: () => void;
   /**
    * Retry limit. Default is 10.
    */
@@ -93,78 +61,45 @@ export interface UploadOptions {
    * Passing true/false toggles the global intelligent flow (all parts are chunked and committed).
    * Passing `'fallback'` will only use FII when network conditions may require it (only failing parts will be chunked).
    */
-  intelligent?: boolean | string;
+  intelligent?: boolean | 'fallback';
   /**
    * Set the default intiial chunk size for Intelligent Ingestion. Defaults to 8MB on desktop and 1MB on mobile.
    */
   intelligentChunkSize?: number;
 }
 
+export type StoreUploadOptions = StoreBaseParams & {
+  /**
+   * Filename or function that returns custom filename for stored file
+   */
+  filename?: ((file: File) => string) | string;
+
+  /**
+   * Workflows ids to run after upload
+   */
+  workflows?: (string | WorkflowConfig)[];
+
+  /**
+   * Fielname sanitizer for cleanup  before upload
+   *
+   * @see https://docs.aws.amazon.com/AmazonS3/latest/dev/UsingMetadata.html
+   * @default {
+   *   exclude: ['\\', '{', '}','|', '%', '`', '"', "'", '~', '[', ']', '#', '|', '^', '<', '>']
+   *   replacement: '-'
+   * }
+   * @type {((boolean | {
+   *     exclude: string[],
+   *     replacement: string,
+   *   }))}
+   */
+  sanitizer?: SanitizeOptions
+};
+
+export interface WorkflowConfig {
+  id: string;
+}
+
 export interface FSProgressEvent {
   totalPercent: number;
   totalBytes: number;
-}
-
-export interface FSRetryEvent {
-  location: string;
-  parts: PartsMap;
-  filename: string;
-  attempt: number | undefined;
-  chunkSize?: number;
-}
-
-/**
- * @private
- */
-export interface UploadConfig extends UploadOptions {
-  apikey: string;
-  store: any;
-  concurrency: number;
-  partSize: number;
-  retryFactor: number;
-  retryMaxTime: number;
-  progressInterval: number;
-  policy?: string;
-  signature?: string;
-  customName?: string;
-  mimetype?: string;
-}
-
-/**
- * @private
- */
-export const enum Status {
-  INIT = 'init',
-  RUNNING = 'running',
-  DONE = 'done',
-  FAILED = 'failed',
-  PAUSED = 'paused',
-}
-
-/**
- * @private
- */
-export interface PartsMap {
-  [part: string]: PartObj;
-}
-
-/**
- * @private
- */
-export interface State {
-  progressTick: any;
-  previousPayload: any;
-  status: Status;
-  retries: any;
-  parts: PartsMap;
-}
-
-/**
- * @private
- */
-export interface Context {
-  config: UploadConfig;
-  state: State;
-  file: FileObj;
-  params?: any;
 }
