@@ -15,14 +15,17 @@
  * limitations under the License.
  */
 
+import * as EventEmitter from 'eventemitter3';
 import { config, Hosts } from '../config';
 import { metadata, MetadataOptions, remove, retrieve, RetrieveOptions } from './api/file';
 import { transform, TransformOptions } from './api/transform';
 import { storeURL } from './api/store';
-import { resolveHost } from './utils/index';
-import { upload, UploadOptions } from './api/upload';
+import { resolveHost } from './utils';
+import { Upload, InputFile, UploadOptions, StoreUploadOptions } from './api/upload';
 import { preview, PreviewOptions } from './api/preview';
 import { CloudClient } from './api/cloud';
+import { StoreParams } from './filelink';
+
 import {
   picker,
   PickerInstance,
@@ -40,42 +43,6 @@ export interface Session {
 export interface Security {
   policy: string;
   signature: string;
-}
-
-export interface WorkflowConfig {
-  id: string;
-}
-
-export interface StoreOptions {
-  /**
-   * Filename for stored file
-   */
-  filename?: string;
-  /**
-   * Location for stored file. One of 's3', 'gcs', 'azure', 'rackspace', or 'dropbox'.
-   */
-  location?: string;
-  /**
-   * Set container path.
-   */
-  path?: string;
-  /**
-   * Specify S3 region.
-   */
-  region?: string;
-  /**
-   * Specify storage container.
-   */
-  container?: string;
-  /**
-   * S3 container access. 'public' or 'private'.
-   */
-  access?: string;
-
-  /**
-   * Workflows ids to run after upload
-   */
-  workflows?: (string | WorkflowConfig)[];
 }
 
 export interface ClientOptions {
@@ -117,37 +84,62 @@ export interface ClientOptions {
  * const client = filestack.init('apikey');
  * ```
  */
-export class Client {
+export class Client extends EventEmitter {
   session: Session;
   private cloud: CloudClient;
 
   constructor(apikey: string, options?: ClientOptions) {
+    super();
+
     if (!apikey || typeof apikey !== 'string' || apikey.length === 0) {
       throw new Error('An apikey is required to initialize the Filestack client');
     }
     const { urls } = config;
     this.session = { apikey, urls };
+
     if (options) {
       const { cname, security } = options;
-      if (security && !(security.policy && security.signature)) {
-        throw new Error('Both policy and signature are required for client security');
-      }
-      if (security && security.policy && security.signature) {
-        this.session.policy = security.policy;
-        this.session.signature = security.signature;
-      }
-      if (cname) {
-        this.session.urls = resolveHost(this.session.urls, cname);
-        const hosts = /filestackapi.com|filestackcontent.com/i;
-        this.session.cname = cname;
-        Object.keys(urls).forEach((key) => {
-          this.session.urls[key] = urls[key].replace(hosts, cname);
-        });
-      }
+
+      this.setSecurity(security);
+      this.setCname(cname);
     }
 
     this.cloud = new CloudClient(this.session, options);
   }
+
+  /**
+   * Set security object
+   *
+   * @param {Security} security
+   * @memberof Client
+   */
+  setSecurity(security: Security) {
+    if (security && !(security.policy && security.signature)) {
+      throw new Error('Both policy and signature are required for client security');
+    }
+
+    if (security && security.policy && security.signature) {
+      this.session.policy = security.policy;
+      this.session.signature = security.signature;
+    }
+  }
+
+  /**
+   * Set custom cname
+   *
+   * @param {string} cname
+   * @returns
+   * @memberof Client
+   */
+  setCname(cname: string) {
+    if (!cname || cname.length === 0) {
+      return;
+    }
+
+    this.session.cname = cname;
+    this.session.urls = resolveHost(this.session.urls, cname);
+  }
+
   /**
    * Clear all current cloud sessions in the picker.
    * Optionally pass a cloud source name to only log out of that cloud source.
@@ -272,7 +264,7 @@ export class Client {
    * @param token     Optional control token to call .cancel()
    * @param security  Optional security override.
    */
-  storeURL(url: string, options?: StoreOptions, token?: any, security?: Security): Promise<Object> {
+  storeURL(url: string, options?: StoreParams, token?: any, security?: Security): Promise<Object> {
     /* istanbul ignore next */
     return storeURL(this.session, url, options, token, security);
   }
@@ -363,11 +355,14 @@ export class Client {
    * client.upload(file, { onRetry }, { filename: 'foobar.jpg' }, token)
    *   .then(res => console.log(res));
    *
+   * client.upload({file, name}, { onRetry }, { filename: 'foobar.jpg' }, token)
+   *   .then(res => console.log(res));
+   *
    * token.pause();  // Pause flow
    * token.resume(); // Resume flow
    * token.cancel(); // Cancel flow (rejects)
    * ```
-   * @param file           Must be a valid [File](https://developer.mozilla.org/en-US/docs/Web/API/File), Blob, base64 encoded string, or file path in Node.
+   * @param file           Must be a valid [File | Blob | Buffer | string]
    * @param uploadOptions  Uploader options.
    * @param storeOptions   Storage options.
    * @param token          A control token that can be used to call cancel(), pause(), and resume().
@@ -375,8 +370,74 @@ export class Client {
    *
    * @returns {Promise}
    */
-  upload(file: any, options?: UploadOptions, storeOptions?: StoreOptions, token?: any, security?: Security) {
-    /* istanbul ignore next */
-    return upload(this.session, file, options, storeOptions, token, security);
+  upload(file: InputFile, options?: UploadOptions, storeOptions?: StoreUploadOptions, token?: any, security?: Security) {
+    let upload = new Upload(options, storeOptions);
+    upload.setSession(this.session);
+
+    if (token) {
+      upload.setToken(token);
+    }
+
+    if (security) {
+      upload.setSecurity(security);
+    }
+
+    upload.on('error', (e) => this.emit('uploadError', e));
+
+    return upload.upload(file).finally(() => {
+      upload = null;
+    });
+  }
+
+  /**
+   * Initiates a multi-part upload flow. Use this for Filestack CIN and FII uploads.
+   *
+   * In Node runtimes the file argument is treated as a file path.
+   * Uploading from a Node buffer is not yet implemented.
+   *
+   * ### Example
+   *
+   * ```js
+   * const token = {};
+   * const onRetry = (obj) => {
+   *   console.log(`Retrying ${obj.location} for ${obj.filename}. Attempt ${obj.attempt} of 10.`);
+   * };
+   *
+   * client.multiupload([file], { onRetry }, token)
+   *   .then(res => console.log(res));
+   *
+   * client.multiupload([{file, name}], { onRetry }, token)
+   *   .then(res => console.log(res));
+   *
+   * token.pause();  // Pause flow
+   * token.resume(); // Resume flow
+   * token.cancel(); // Cancel flow (rejects)
+   * ```
+   * @param file           Must be a valid [File | Blob | Buffer | string (base64)]
+   * @param uploadOptions  Upload options.
+   * @param storeOptions   Storage options.
+   * @param token          A control token that can be used to call cancel(), pause(), and resume().
+   * @param security       Optional security policy and signature override.
+   *
+   * @returns {Promise}
+   */
+  multiupload(file: InputFile[], options?: UploadOptions, storeOptions?: StoreUploadOptions, token?: any, security?: Security) {
+    let upload = new Upload(options, storeOptions);
+
+    upload.setSession(this.session);
+
+    if (token) {
+      upload.setToken(token);
+    }
+
+    if (security) {
+      upload.setSecurity(security);
+    }
+
+    upload.on('error', (e) => this.emit('uploadError', e));
+
+    return upload.multiupload(file).finally(() => {
+      upload = null;
+    });
   }
 }
