@@ -16,8 +16,10 @@
  */
 
 import { loadModule, knownModuleIds } from 'filestack-loader';
-import { Client, WorkflowConfig } from './client';
-import { FSProgressEvent, UploadOptions } from './api/upload/types';
+import { FilestackError, FilestackErrorType } from './../filestack_error';
+import { Client } from './client';
+import { FSProgressEvent, UploadOptions, WorkflowConfig } from './api/upload/types';
+import { getValidator, PickerParamsSchema } from './../schema';
 
 export interface PickerInstance {
   /**
@@ -261,88 +263,89 @@ export interface PickerStoreOptions {
 
 export interface PickerCustomText {
   // Actions
-  Upload: string;
-  'Deselect All': string;
-  'View/Edit Selected': string;
-  'Sign Out': string;
+  Upload?: string;
+  'Deselect All'?: string;
+  'View/Edit Selected'?: string;
+  'Sign Out'?: string;
 
   // Source Labels
-  'My Device': string;
-  'Web Search': string;
-  'Take Photo': string;
-  'Link (URL)': string;
-  'Record Video': string;
-  'Record Audio': string;
+  'My Device'?: string;
+  'Web Search'?: string;
+  'Take Photo'?: string;
+  'Link (URL)'?: string;
+  'Record Video'?: string;
+  'Record Audio'?: string;
 
   // Custom Source
-  'Custom Source': string;
+  'Custom Source'?: string;
 
   // Footer Text
-  Add: string;
-  'more file': string;
-  'more files': string;
+  Add?: string;
+  'more file'?: string;
+  'more files'?: string;
 
   // Cloud
-  Connect: string;
-  'Select Files from': string;
-  'You need to authenticate with ': string;
-  'A new page will open to connect your account.': string;
-  'We only extract images and never modify or delete them.': string;
+  Connect?: string;
+  'Select Files from'?: string;
+  'You need to authenticate with '?: string;
+  'A new page will open to connect your account.'?: string;
+  'We only extract images and never modify or delete them.'?: string;
 
   // Summary
-  Files: string;
-  Images: string;
-  Uploaded: string;
-  Uploading: string;
-  Completed: string;
-  Filter: string;
-  'Cropped Images': string;
-  'Edited Images': string;
-  'Selected Files': string;
-  'Crop is required on images': string;
+  Files?: string;
+  Images?: string;
+  Uploaded?: string;
+  Uploading?: string;
+  Completed?: string;
+  Filter?: string;
+  'Cropped Images'?: string;
+  'Edited Images'?: string;
+  'Selected Files'?: string;
+  'Crop is required on images'?: string;
 
   // Transform
-  Crop: string;
-  Circle: string;
-  Rotate: string;
-  Mask: string;
-  Revert: string;
-  Edit: string;
-  Reset: string;
-  Done: string;
-  Save: string;
-  Next: string;
-  'Edit Image': string;
-  'This image cannot be edited': string;
+  Crop?: string;
+  Circle?: string;
+  Rotate?: string;
+  Mask?: string;
+  Revert?: string;
+  Edit?: string;
+  Reset?: string;
+  Done?: string;
+  Save?: string;
+  Next?: string;
+  'Edit Image'?: string;
+  'This image cannot be edited'?: string;
 
   // Retry messaging
-  'Connection Lost': string;
-  'Failed While Uploading': string;
-  'Retrying in': string;
-  'Try again': string;
-  'Try now': string;
+  'Connection Lost'?: string;
+  'Failed While Uploading'?: string;
+  'Retrying in'?: string;
+  'Try again'?: string;
+  'Try now'?: string;
 
   // Local File Source
-  'or Drag and Drop, Copy and Paste Files': string;
-  'Select Files to Upload': string;
-  'Select From': string;
-  'Drop your files anywhere': string;
+  'or Drag and Drop, Copy and Paste Files'?: string;
+  'Select Files to Upload'?: string;
+  'Select From'?: string;
+  'Drop your files anywhere'?: string;
 
   // Input placeholders
-  'Enter a URL': string;
-  'Search images': string;
+  'Enter a URL'?: string;
+  'Search images'?: string;
 
   // Webcam Source
-  'Webcam Disabled': string;
-  'Webcam Not Supported': string;
-  'Please enable your webcam to take a photo.': string;
-  'Your current browser does not support webcam functionality.': string;
-  'We suggest using Chrome or Firefox.': string;
+  'Webcam Disabled'?: string;
+  'Webcam Not Supported'?: string;
+  'Please enable your webcam to take a photo.'?: string;
+  'Your current browser does not support webcam functionality.'?: string;
+  'We suggest using Chrome or Firefox.'?: string;
 
   // Error Notifications
-  'File {displayName} is not an accepted file type. The accepted file types are {types}': string;
-  'File {displayName} is too big. The accepted file size is less than {roundFileSize}': string;
-  'Our file upload limit is {maxFiles} {filesText}': string;
+  'File {displayName} is not an accepted file type. The accepted file types are {types}'?: string;
+  'File {displayName} is too big. The accepted file size is less than {roundFileSize}'?: string;
+  'Our file upload limit is {maxFiles} {filesText}'?: string;
+  'No search results found for "{search}"'?: string;
 }
 
 export interface PickerOptions {
@@ -513,6 +516,11 @@ export interface PickerOptions {
    * Restrict selected files to a maximum number of bytes. (e.g. 10 \* 1024 \* 1024 for 10MB limit).
    */
   maxSize?: number;
+
+  /**
+   * Default view type option for file browser
+   */
+  viewType?: 'grid' | 'list';
   /**
    * Specify [width, height] in pixels of the desktop modal.
    */
@@ -591,10 +599,6 @@ export interface PickerOptions {
    * Called when all files have been uploaded.
    */
   onUploadDone?: PickerUploadDoneCallback;
-  /**
-   * For cloud sources whether to link or store files. Defaults to `false`.
-   */
-  preferLinkOverStore?: boolean;
 
   /**
    * Define a unique id for the application mount point.
@@ -656,6 +660,11 @@ export interface PickerTransformationOptions {
    * Enable image rotation. Defaults to `true`.
    */
   rotate?: boolean;
+
+  /**
+   * Global force crop option. Can be use ie with circle
+   */
+  force?: boolean;
 }
 
 /**
@@ -669,6 +678,11 @@ class PickerLoader {
   private _initialized: Promise<PickerInstance>;
 
   constructor(client: Client, options?: PickerOptions) {
+    const validateRes = getValidator(PickerParamsSchema)(options);
+    if (validateRes.errors.length) {
+      throw new FilestackError(`Invalid picker params`, validateRes.errors, FilestackErrorType.VALIDATION);
+    }
+
     this._initialized = this.loadModule(client, options);
   }
 
